requirements.txt
pandas
openpyxl
spidev
rpi.gpio
smbus
python-can
numpy
tk
psutil
setuptools